Consolidate 2 json.dart files into 1. make new GithubJson part of common.
