+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 6.2.0
+ - Bug fix: Fix setRepositorySubscription to be a PUT instead of a POST https://github.com/SpinlockLabs/github.dart/commit/5b5d7656ce9ce1cb06e15651da06e7e192bc19e1
+ - Bug fix: Repository clone URLs were null. DEPRECATED `Repository.cloneUrls` use `cloneUrl`,`gitUrl`,`sshUrl`, or `svnUrl` instead.
+ - Bug fix: Use a shared json encoder util to remove nulls from maps and lists, encode all dates for github. https://github.com/SpinlockLabs/github.dart/pull/182
+
 ## 6.1.3
  - Add missing fields for Notification https://github.com/SpinlockLabs/github.dart/pull/210
  - Can now create draft PRs https://github.com/SpinlockLabs/github.dart/pull/212
