+++ lib/src/common/issues_service.dart
@@ -4,7 +4,6 @@ import 'dart:convert';
 import 'package:github/src/common.dart';
 import 'package:github/src/common/model/users.dart';
 import 'package:github/src/common/util/pagination.dart';
-import 'package:github/src/json.dart';
 
 /// The [IssuesService] handles communication with issues related methods of the
 /// GitHub API.
