+++ lib/src/common/pulls_service.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:convert';
 import 'package:github/src/common.dart';
 import 'package:github/src/common/util/pagination.dart';
-import 'package:github/src/json.dart';
 import 'package:github/src/util.dart';
 
 /// The [PullRequestsService] handles communication with pull request
